requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg==3.1.16
psycopg-binary==3.1.16
PyJWT==2.8.0
python-dotenv==1.0.0
typing_extensions==4.9.0
Werkzeug==3.0.1
